+++ internal/services/v1/experimental_test.go
@@ -157,7 +157,7 @@ func TestBulkExportRelationshipsBeyondAllowedLimit(t *testing.T) {
 
 	_, err = resp.Recv()
 	require.Error(err)
-	require.Contains(err.Error(), "provided limit 10000005 is greater than maximum allowed of 100000")
+	require.Contains(err.Error(), "provided limit 10000005 is greater than maximum allowed of 10000")
 }
 
 func TestBulkExportRelationships(t *testing.T) {